requirements.txt
google-generativeai
pywin32
openpyxl
python-dotenv

